GridScanner/requirements.txt
fastapi==0.60.1
Jinja2==2.11.2
uvicorn==0.11.8
aiofiles==0.5.0
python-multipart==0.0.5
pdf2image==1.13.1
requests==2.21.0
numpy==1.15.4
opencv-python==4.0.0.21
tensorflow==1.12
scipy==1.2.0
tensorflow-serving-api==1.12.0
Pillow==5.4.0
protobuf==3.6.1